Organize controller logic into packages
